Extract brkitr/root.txt changes to patches
There is a conflict in this file when upgrading to ICU 77.1, and having
it in data_files_to_preserve.txt makes it harder to see what the changes
are, or to review changes to the implicit patch.

The last time the file was modified was in the upgrade to ICU 63.1:
https://chromium.googlesource.com/chromium/deps/icu/+/42d5027992a0946942839b8821765e1512afbc21

At the time the only difference was the line_normal change, compare ICU
63.1 to our version:
https://github.com/unicode-org/icu/blob/release-63-1/icu4c/source/data/brkitr/root.txt
https://chromium.googlesource.com/chromium/deps/icu/+/42d5027992a0946942839b8821765e1512afbc21/source/data/brkitr/root.txt

That difference is extracted into line_normal.patch.

Since then an lstm section has been added in ICU, which has been
reverted by the data_files_to_preserve.txt mechanism.

Adding the lstm section changes the built output of make_data_all.sh, so
add a separate patch to remove it to not change anything with this CL.

Bug: 421834885
Change-Id: I08098a68b32d56395c4798d746029e0cfc4831ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/deps/icu/+/6788296
Reviewed-by: Koji Ishii <kojii@chromium.org>
Reviewed-by: David Yeung <dayeung@chromium.org>